api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
requests==2.31.0
pydantic-settings==2.1.0
httpx==0.26.0
pytest==8.0.0
